+++ client/src/App.jsx
@@ -1,0 +1,41 @@
+import {BrowserRouter as Router, Route, Routes} from 'react-router-dom';
+import './App.css';
+import {routers} from './routers/router';
+import Header from './partials/Header';
+import Footer from './partials/Footer';
+import React, {Suspense, useState} from 'react';
+
+
+function App() {
+  const [formData, setFormData] = useState();
+  return (
+      <div className="App">
+        <Suspense fallback="Loading...">
+        </Suspense>
+        <Router>
+          <Routes>
+            {routers.map((r) => {
+              const Page = r.page;
+              const Layout = r.isShowHeader ? <Header/> : <></>;
+              const FooterLayout = r.isShowFooter ? <Footer/> : <></>;
+              return (
+                  <Route
+                      key={r.path}
+                      path={r.path}
+                      element={
+                        <>
+                          {Layout}
+                          <Page/>
+                          {FooterLayout}
+                        </>
+                      }
+                  />
+              );
+            })}
+          </Routes>
+        </Router>
+      </div>
+  );
+}
+
+export default App;
