+++ client/src/index.jsx
@@ -1,13 +1,13 @@
-import React from 'react';
+import React, {StrictMode} from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
+    <StrictMode>
+      <App/>
+    </StrictMode>,
 );
 
 // If you want to start measuring performance in your app, pass a function
