+++ client/src/partials/Header.jsx
@@ -69,8 +69,8 @@ const Header = ({ children }) => {
       <Row justify={"center"} className="bg-[#114098]">
         <Col span={2}></Col>
         <Col span={20}>
-          <div class="bg-[#114098] flex justify-between items-center pt-4 pb-1 relative">
-            <ul class="flex pt-7 mt-3 ">
+          <div className="bg-[#114098] flex justify-between items-center pt-4 pb-1 relative">
+            <ul className="flex pt-7 mt-3 ">
               <div className="absolute top-2 text-white left-[3%] text-[25px] font-lobster cursor-pointer pt-2">
                 {" "}
                 Take A Breath
@@ -98,7 +98,7 @@ const Header = ({ children }) => {
                 </Link>
               </li>
             </ul>
-            <ul class="flex space-x-5 pt-3">
+            <ul className="flex space-x-5 pt-3">
               <li>
                 <Link to="/" className="no-underline">
                   <Button>Log In</Button>
